backend/requirements.txt
asgiref==3.9.1
django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
